refactor chapter rules
